+++ Dockerfile
@@ -1,10 +1,6 @@
-FROM node:10
+FROM node:14
 WORKDIR /usr/src/app
-
-COPY package.json .
-
-RUN npm install
-
 COPY . .
-
+RUN npm install
+RUN npm run build
 CMD [ "npm", "start" ]
