+++ package.json
@@ -4,7 +4,10 @@
   "description": "node_service_eventstream-api",
   "main": "app.js",
   "scripts": {
-    "start": "node app.js"
+    "dev": "nodemon --exec babel-node src/app.js",
+    "build": "babel src --out-dir dist",
+    "start": "node dist/app.js",
+    "lint": "eslint src/"
   },
   "dependencies": {
     "cookie-parser": "^1.4.5",
@@ -15,7 +18,6 @@
     "httpntlm": "^1.7.7",
     "ladc": "^0.22.14",
     "md5": "^2.3.0",
-    "morgan": "^1.10.0",
     "pg": "^8.5.1",
     "pug": "3.0.2",
     "sequelize": "^6.5.0",
@@ -32,5 +34,13 @@
   "bugs": {
     "url": "https://github.com/StadGent/node_service_eventstream-api/issues"
   },
-  "homepage": "https://github.com/StadGent/node_service_eventstream-api#readme"
+  "homepage": "https://github.com/StadGent/node_service_eventstream-api#readme",
+  "devDependencies": {
+    "@babel/cli": "^7.14.3",
+    "@babel/core": "^7.14.3",
+    "@babel/node": "^7.14.2",
+    "@babel/preset-env": "^7.14.2",
+    "eslint": "^7.27.0",
+    "nodemon": "^2.0.7"
+  }
 }
