+++ src/config/config.js
@@ -1,7 +1,7 @@
-let fs = require("fs");
-let toml = require('toml');
-
-module.exports.getConfig = function() {
+import fs from 'fs';
+import toml from 'toml';
+export default class Config {
+  static getConfig() {
     let data = toml.parse(fs.readFileSync(__dirname + "/../config.tml").toString());
     return {
         database: {
@@ -38,13 +38,13 @@ module.exports.getConfig = function() {
             institutionURI: data.stam.institutionURI
         }
     };
-};
-
-module.exports.getInstitutionFromURI = function(institutionURI) {
+  }
+  static getInstitutionFromURI(institutionURI) {
     let config = this.getConfig();
     for(let i in Object.keys(config)) {
-        let key = Object.keys(config)[i];
-        if (config[key].institutionURI && config[key].institutionURI === institutionURI) return key;
+      let key = Object.keys(config)[i];
+      if (config[key].institutionURI && config[key].institutionURI === institutionURI) return key;
     }
     return "";
+  }
 }
